Prometheus monitor alertmanager
We have alerts logic for alerting if the alertmanager gets out of sync,
but this could never fire as previously we were never monitoring the
alertmanager.

This commit adds a ServiceMonitor so Prometheus monitors metrics coming
from its AlertManager. So now the alerting can fire if the AlertManagers
get out of sync.

How can you alert on the alertmanager if your monitoring doesn't have a
ServiceMonitor to monitor and alert on the AlertManager? :D