Do not schedule fluentd on masters
Fluentd needs to watch the API for namespace changes. When running
fluentd on a master on a multi-master cluster, fluentd crashes. This
happens because we hit the internal load balancer, which won't return
when the destination is the same as the source.

See https://github.com/Azure/acs-engine/blob/master/docs/kubernetes/features.md#private-cluster